Moved the PYI-related functions out to their own module.
